Move config.py into core/
