+++ ledger_bot/clients/reaction_roles_client.py
@@ -7,7 +7,7 @@ import discord
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 
-from ledger_bot.config import Config
+from ledger_bot.core import Config
 from ledger_bot.services import Service
 
 from .extended_client import ExtendedClient
