+++ ledger_bot/core/__init__.py
@@ -1,0 +1,5 @@
+"""Core components."""
+
+from . import config
+
+Config = config.Config
