+++ ledger_bot/message_generators/generate_transaction_status_message.py
@@ -5,7 +5,7 @@
 import logging
 from typing import TYPE_CHECKING, Optional
 
-from ledger_bot.config import Config
+from ledger_bot.core import Config
 from ledger_bot.models import Transaction
 
 if TYPE_CHECKING:
